+++ Setup/UpgradeData.php
@@ -51,7 +51,7 @@ class UpgradeData implements UpgradeDataInterface
         
 
         // add custom Order Attribute
-        if (version_compare($context->getVersion(), '2.1.1', '<')) {
+        if (version_compare($context->getVersion(), '2.2.0', '<')) {
             $eavSetup = $this->eavSetupFactory->create(['setup' => $setup]);
             
 //            $eavSetup->removeAttribute(\Magento\Catalog\Model\Product::ENTITY, 'sc_subscription_plans');