+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ---
 
-## 2.1.1.x
+## 2.2.0.x
 ```
 	* Added possibility to append Safecharge Payment Plan to a product, when create new or edit existing one. Plugin get the payment plans from the pre-created plans of the merchant, by click on a button Get Payment Plans in plugin configuration.
 	* When user have a product with a Payment Plan, only CC is allowed.
+	* Added third option for mandatory dorp-downs in the plugin settings.
 
 ## 2.1.0.x
 ```